requirements.txt
# Web UI for live progress (optional for CLI-only `main.py`).
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
